requirements.txt
texteller
fastapi
uvicorn
python-multipart
pillow
numpy
onnxruntime-gpu
